Optimized mem_copy

+++ src/kernel/memory.h
@@ -3,7 +3,7 @@
 #include "types.h"
 
 void mem_set(void *ptr, unsigned char value, int size);
-void mem_copy(void *dst, const void *src, int size);
+void mem_copy(void *dst, const void *src, size_t size);
 
 void *mem_alloc(size_t size);
 void *mem_alloc_zero(size_t size);
